Exceptions Handled

+++ order-service/src/main/java/com/catalogue/orderservice/OrderController.java
@@ -1,6 +1,9 @@
 package com.catalogue.orderservice;
+import org.apache.commons.lang.ObjectUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,29 +18,42 @@ public class OrderController {
 
     @RequestMapping("/orders")
     public List<CatalogueOrder> getAllOrders(){
-        orderRepository.save(new CatalogueOrder("abhiyad1","1",Boolean.FALSE) );
-        orderRepository.save(new CatalogueOrder("abhiyad2","2",Boolean.FALSE) );
         List<CatalogueOrder> order = new ArrayList<>();
         orderRepository.findAll().forEach(order::add);
         return order;
     }
     @RequestMapping("/orders/{id}")
-    public CatalogueOrder getOrder(@PathVariable("id") String id){
-        return orderRepository.findById(Long.valueOf(id)).orElse(null);
+    public CatalogueOrder getOrder(@PathVariable("id") String id)throws ResponseStatusException{
+        if ( orderRepository.existsById(Long.valueOf(id)) )
+            return orderRepository.findById(Long.valueOf(id)).orElse(new CatalogueOrder());
+        else{
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Order not found");
+        }
     }
-    @PostMapping("/orders/update/{id}")
-    public CatalogueOrder update(@PathVariable("id") String id, @RequestBody CatalogueOrder order){
-        order.setId(Long.valueOf(id));
-        orderRepository.save(order);
-        return order;
+    @RequestMapping("/orders/update/{id}")
+    public CatalogueOrder update(@PathVariable("id") String id, @RequestBody CatalogueOrder order)throws ResponseStatusException{
+        if ( orderRepository.existsById(Long.valueOf(id)) ) {
+            order.setId(Long.valueOf(id));
+            orderRepository.save(order);
+            return order;
+        }
+        else{
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Order not found");
+        }
     }
-    @PostMapping("orders/create")
+    @RequestMapping("orders/create")
     public CatalogueOrder create(@RequestBody CatalogueOrder order){
+        order.setId(null);
         orderRepository.save(order);
         return order;
     }
     @RequestMapping("orders/delete/{id}")
-    public void delete(@PathVariable("id") String id){
-        orderRepository.delete(orderRepository.findById(Long.valueOf(id)).orElse(null));
+    public void delete(@PathVariable("id") String id)throws ResponseStatusException{
+        if ( orderRepository.existsById(Long.valueOf(id)) ) {
+            orderRepository.delete(orderRepository.findById(Long.valueOf(id)).orElse(null));
+        }
+        else{
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Order not found");
+        }
     }
 }
